+++ R/count-tally.R
@@ -117,7 +117,7 @@ count <- function(x, ..., wt = NULL, sort = FALSE) {
 #' @export
 #' @rdname se-deprecated
 count_ <- function(x, vars, wt = NULL, sort = FALSE) {
-  vars <- compat_lazy_dots(vars, caller_env(), .named = TRUE)
+  vars <- compat_lazy_dots(vars, caller_env())
   count(x, !!! vars, wt = wt, sort = sort)
 }
 
@@ -170,6 +170,6 @@ add_count <- function(x, ..., wt = NULL, sort = FALSE) {
 #' @rdname se-deprecated
 #' @export
 add_count_ <- function(x, vars, wt = NULL, sort = FALSE) {
-  vars <- compat_lazy_dots(vars, caller_env(), .named = TRUE)
+  vars <- compat_lazy_dots(vars, caller_env())
   add_count(x, !!! vars, wt = wt, sort = sort)
 }