+++ tests/testthat/test-binds.R
@@ -11,6 +11,24 @@ test_that("cbind and rbind methods work for tbl_df", {
   expect_equal(rbind(df1, df2), tibble(x = c(1, NA), y = c(NA, 2)))
 })
 
+
+# error -------------------------------------------------------------------
+
+test_that("bind_rows() and bind_cols() err for non-data frames (#2373)", {
+  df1 <- memdb_frame(x = 1)
+  df2 <- memdb_frame(y = 2)
+
+  expect_error(
+    bind_cols(df1, df2),
+    "Data-frame-like objects must inherit from class data.frame or be plain lists"
+  )
+  expect_error(
+    bind_rows(df1, df2),
+    "Data-frame-like objects must inherit from class data.frame or be plain lists"
+  )
+})
+
+
 # columns -----------------------------------------------------------------
 
 test_that("cbind uses shallow copies", {
@@ -128,23 +146,31 @@ test_that("bind_rows handles data frames with no columns (#1346)", {
   expect_equal(dim(bind_rows(df0, df0)), c(2, 0))
 
   res <- bind_rows(df0, df1)
-  expect_equal(res$x, c(1, NA))
+  expect_equal(res$x, c(NA, 1))
 })
 
 test_that("bind_rows handles lists with NULL values (#2056)", {
-  skip("Currently failing")
   df1 <- data_frame(x = 1, y = 1)
   df2 <- data_frame(x = 2, y = 2)
   lst1 <- list(a = df1, NULL, b = df2)
-  lst2 <- list(a = df1, b = df2)
 
   df3 <- data_frame(
     names = c("a", "b"),
-    x = 1:2,
-    y = 1:2
+    x = c(1, 2),
+    y = c(1, 2)
   )
 
   expect_equal(bind_rows(lst1, .id = "names"), df3)
+})
+
+test_that("bind_rows puts data frames in order received even if no columns (#2175)", {
+  df2 <- data_frame(x = 2, y = "b")
+  df1 <- df2[, 0]
+
+  res <- bind_rows(df1, df2)
+
+  expect_equal(res$x, c(NA, 2))
+  expect_equal(res$y, c(NA, "b"))
 })
 
 # Column coercion --------------------------------------------------------------